backend/requirements.txt
# OpenLedger Black - Backend Dependencies
# Zero external paid services, enterprise-grade stack

# FastAPI - Lightning-fast async web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-auth>=2.23.0
slowapi>=0.1.9

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client for external APIs
httpx>=0.24.1

# PDF Generation
reportlab==4.0.9

# Excel/CSV
openpyxl==3.1.2

# Date handling
python-dateutil==2.8.2

# ML and Analytics
scikit-learn>=1.3.0
statsmodels>=0.14.0
joblib>=1.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
